Payment

+++ src/app/user/home/payment/payment.component.ts
@@ -4,7 +4,9 @@ import {Category} from "../../model/category";
 import {PaymentService} from "../../service/payment.service";
 import {CategoryService} from "../../service/category.service";
 import {Payment} from "../../model/payment";
-
+import {WalletService} from "../../service/wallet.service";
+import Swal from "sweetalert2";
+import {Wallet} from "../../model/wallet";
 
 @Component({
   selector: 'app-payment',
@@ -35,20 +37,34 @@ export class PaymentComponent implements OnInit {
   nameCategory: string = 'Danh mục giao dich';
   expenseCategories: Category[] = [];
   incomeCategories: Category[] = [];
+  listWallet: Wallet[] = [];
+  wallet_id: number = 0;
 
   constructor(
     private paymentService: PaymentService,
-    private categoryService: CategoryService
+    private categoryService: CategoryService,
+    private walletService: WalletService,
   ) {
     this.categoryService.findAll().subscribe(data => {
       this.listCategory = data;
       console.log("cTWA", this.listCategory);
     })
+    this.walletService.getAll().subscribe(data => {
+      this.listWallet = data;
+      console.log("wallet", this.listWallet);
+      this.wallet_id = this.listWallet[0].id;
+      console.log("wallet_id", this.wallet_id);
+      this.getPaymentList();
+    })
+
   }
 
   ngOnInit(): void {
-    this.getPaymentList();
+  }
 
+  getWalletid(id: any) {
+    this.wallet_id = id;
+    console.log(this.wallet_id);
   }
 
 
@@ -57,13 +73,22 @@ export class PaymentComponent implements OnInit {
     console.log(data);
     console.log(data.category);
     if (data.date == null) {
+      Swal.fire('Fail', 'Please do not blank', 'error');
+
       data.date = new Date();
       console.log(data.date);
     }
     data.category = {
       id: data.category
     };
+
+    data.wallet = {
+      id: this.wallet_id
+    }
     this.paymentService.save(data).subscribe(data => {
+      Swal.fire('Sucess',
+        'Please choose date',
+        'success');
       console.log(data)
       this.paymentForm.reset();
       this.getPaymentList();
@@ -73,17 +98,42 @@ export class PaymentComponent implements OnInit {
   paymentList: Payment[] = [];
 
   getPaymentList() {
-    this.paymentService.findAll().subscribe(data => {
+    this.paymentService.findAllByWallet(this.wallet_id).subscribe(data => {
       this.paymentList = data;
       console.log("paymentList", this.paymentList);
     })
   }
 
+
   deletePayment(id: number) {
-    this.paymentService.delete(id).subscribe(data => {
-      this.getPaymentList();
+    Swal.fire({
+      title: 'Are You Sure?',
+      text: "You Won't Be Able To Revert This!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'No Keep It',
+      confirmButtonText: 'Yes, Delete It!',
     })
+      .then((result) => {
+        if (result.isConfirmed) {
+          this.paymentService.delete(id).subscribe(data => {
+            this.getPaymentList();
+          }, e => {
+            console.log(e);
+          });
+          Swal.fire(
+            'Deleted!',
+            'Your Payment Has Been Deleted.',
+            'success'
+          )
+        }
+      })
+
+
   }
+
 
   // Update payment
   updatePayment(id: number) {
@@ -114,6 +164,9 @@ export class PaymentComponent implements OnInit {
     };
     this.paymentService.update(data.id, data).subscribe(data => {
       console.log(data)
+      Swal.fire('Success',
+        'You Update Your Wallet Successful',
+        'success');
       this.paymentForm.reset();
       this.getPaymentList();
     });
@@ -129,3 +182,13 @@ export class PaymentComponent implements OnInit {
     })
   }
 }
+
+
+
+
+
+
+
+
+
+
